Store money as exact whole Rupiah instead of rounded thousands
Every *_idr amount was silently rounded to the nearest Rp 1,000 on
save (integer "thousands of IDR" convention). Removes that rounding
end-to-end: migration rescales existing data x1000, the backend
stops dividing/rounding to thousands, and the frontend gets a new
exact formatter for every page except the Dashboard summary and its
persistent sidebar total, which keep the abbreviated "Rp3.70 B" style.

+++ backend/internal/httpapi/holdings.go
@@ -18,7 +18,7 @@ type holdingRequest struct {
 	Brand      *string  `json:"brand"`
 	UsdValue   *float64 `json:"usd_value"`
 	Currency   *string  `json:"currency"`
-	ValueIdr   *float64 `json:"value_idr"`
+	ValueIdr   *int64   `json:"value_idr"`
 	Detail     *string  `json:"detail"`
 }
